// product-service/src/core/utils/ErrorCode.enum.ts
const enum Message {
  CATEGORY_ALREADY_EXISTS = "PRODUCT_ALREADY_EXISTS",
  CATEGORY_NOT_FOUND = "PRODUCT_NOT_FOUND",
  CATEGORY_TOO_MANY_ATTEMPTS = "PRODUCT_TOO_MANY_ATTEMPTS",
 CATEGORY_CREATED_SUCCESFULLY= "PRODUCT_CREATED_SUCCESFULLY",
 CATEGORY_UPDATED_SUCCESFULLY= "PRODUCT_UPDATED_SUCCESFULLY",
 CATEGORY_FETCH_SUCCESSFULLY = "PRODUCT_FETCH_SUCCESSFULLY",
  CATEGORY_DELETED_SUCCESSFULLY = "PRODUCT_DELETED_SUCCESSFULLY",
  NO_CATEGORY_FOUND = "NO_PRODUCT_FOUND",
  // Access Control Errors
  ACCESS_FORBIDDEN = "ACCESS_FORBIDDEN",
  ACCESS_UNAUTHORIZED = "ACCESS_UNAUTHORIZED",

  // Validation and Resource Errors
  VALIDATION_ERROR = "VALIDATION_ERROR",
  RESOURCE_NOT_FOUND = "RESOURCE_NOT_FOUND",
  BAD_REQUEST = "BAD REQUEST",

  // System Errors
  INTERNAL_SERVER_ERROR = "INTERNAL_SERVER_ERROR",
  VERIFICATION_ERROR = "VERIFICATION_ERROR",
}

export { Message };